+++ packages/pluggableWidgets/datagrid-date-filter-web/src/DatagridDateFilter.tsx
@@ -8,8 +8,30 @@ import * as locales from "date-fns/locale";
 import { getFilterDispatcher } from "./utils/provider";
 import { Alert } from "@mendix/piw-utils-internal";
 
+import { DefaultFilterEnum } from "../../datagrid-text-filter-web/typings/DatagridTextFilterProps";
+import { chanteTimeToMidnight } from "./utils/utils";
+import { addDays, subDays } from "date-fns";
+
+import {
+    and,
+    attribute as attributeFunction,
+    greaterThan,
+    greaterThanOrEqual,
+    lessThan,
+    lessThanOrEqual,
+    literal,
+    or
+} from "mendix/filters/builders";
+import { FilterCondition } from "mendix/filters";
+
 interface Locale {
     [key: string]: object;
+}
+
+function getAttributeTypeErrorMessage(type?: string): string | null {
+    return type && type !== "DateTime"
+        ? "The attribute type being used for Data grid date filter is not 'Date and time'"
+        : null;
 }
 
 export default function DatagridDateFilter(props: DatagridDateFilterContainerProps): ReactElement | null {
@@ -34,15 +56,62 @@ export default function DatagridDateFilter(props: DatagridDateFilterContainerPro
 
     return FilterContext?.Consumer ? (
         <FilterContext.Consumer>
-            {({ filterDispatcher, attribute }) =>
-                filterDispatcher ? (
+            {filterContextValue => {
+                if (!filterContextValue || !filterContextValue.filterDispatcher || !filterContextValue.attribute) {
+                    return alertMessage;
+                }
+                const { filterDispatcher, attribute } = filterContextValue;
+
+                const errorMessage = getAttributeTypeErrorMessage(attribute.type);
+                if (errorMessage) {
+                    return <Alert bootstrapStyle="danger">{errorMessage}</Alert>;
+                }
+
+                const getFilterConditions = (
+                    value: Date | null,
+                    type: DefaultFilterEnum
+                ): FilterCondition | undefined => {
+                    if (!attribute || !attribute.filterable || !value) {
+                        return undefined;
+                    }
+
+                    const filterAttribute = attributeFunction(attribute.id);
+
+                    switch (type) {
+                        case "greater":
+                            // > Date at midnight
+                            return greaterThan(filterAttribute, literal(chanteTimeToMidnight(value)));
+                        case "greaterEqual":
+                            // > day -1 at midnight
+                            return greaterThan(filterAttribute, literal(subDays(chanteTimeToMidnight(value), 1)));
+                        case "equal":
+                            // >= day at midnight and < day +1 midnight
+                            return and(
+                                greaterThanOrEqual(filterAttribute, literal(chanteTimeToMidnight(value))),
+                                lessThan(filterAttribute, literal(addDays(chanteTimeToMidnight(value), 1)))
+                            );
+                        case "notEqual":
+                            // < day at midnight or >= day +1 at midnight
+                            return or(
+                                lessThan(filterAttribute, literal(chanteTimeToMidnight(value))),
+                                greaterThanOrEqual(filterAttribute, literal(addDays(chanteTimeToMidnight(value), 1)))
+                            );
+                        case "smaller":
+                            // <= day -1 at midnight
+                            return lessThanOrEqual(filterAttribute, literal(subDays(chanteTimeToMidnight(value), 1)));
+                        case "smallerEqual":
+                            // < day +1 at midnight
+                            return lessThan(filterAttribute, literal(addDays(chanteTimeToMidnight(value), 1)));
+                    }
+                };
+                return (
                     <FilterComponent
                         adjustable={props.adjustable}
-                        attribute={attribute}
                         defaultFilter={props.defaultFilter}
                         defaultValue={props.defaultValue?.value}
                         dateFormat={patterns.date}
                         filterDispatcher={filterDispatcher}
+                        getFilterConditions={getFilterConditions}
                         locale={language}
                         name={props.name}
                         placeholder={props.placeholder?.value}
@@ -50,10 +119,8 @@ export default function DatagridDateFilter(props: DatagridDateFilterContainerPro
                         screenReaderInputCaption={props.screenReaderInputCaption?.value}
                         tabIndex={props.tabIndex}
                     />
-                ) : (
-                    alertMessage
-                )
-            }
+                );
+            }}
         </FilterContext.Consumer>
     ) : (
         alertMessage
